format(TextureCache): Moved Texture2D and TextureArray caching and calculation from AssetCache to newly created TextureCache file.
