+++ Assets/GothicVR/Scripts/Caches/TextureCache.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7b0beb2925b74f7caf6f348cd571b599
+timeCreated: 1710684972
